# Conflicts:
#	yudao-module-infra/yudao-module-infra-biz/pom.xml
#	yudao-module-infra/yudao-module-infra-biz/src/main/java/cn/iocoder/yudao/module/infra/controller/admin/codegen/CodegenController.java
#	yudao-module-infra/yudao-module-infra-biz/src/main/java/cn/iocoder/yudao/module/infra/controller/admin/codegen/vo/column/CodegenColumnSaveReqVO.java
#	yudao-module-infra/yudao-module-infra-biz/src/main/java/cn/iocoder/yudao/module/infra/service/codegen/CodegenServiceImpl.java

+++ yudao-module-mall/yudao-module-product-biz/src/main/java/cn/iocoder/yudao/module/product/service/category/ProductCategoryServiceImpl.java
@@ -1,20 +1,27 @@
 package cn.iocoder.yudao.module.product.service.category;
 
+import cn.hutool.core.collection.CollUtil;
 import cn.iocoder.yudao.framework.common.enums.CommonStatusEnum;
+import cn.iocoder.yudao.framework.common.util.collection.CollectionUtils;
 import cn.iocoder.yudao.module.product.controller.admin.category.vo.ProductCategoryCreateReqVO;
 import cn.iocoder.yudao.module.product.controller.admin.category.vo.ProductCategoryListReqVO;
 import cn.iocoder.yudao.module.product.controller.admin.category.vo.ProductCategoryUpdateReqVO;
 import cn.iocoder.yudao.module.product.convert.category.ProductCategoryConvert;
 import cn.iocoder.yudao.module.product.dal.dataobject.category.ProductCategoryDO;
 import cn.iocoder.yudao.module.product.dal.mysql.category.ProductCategoryMapper;
+import cn.iocoder.yudao.module.product.service.spu.ProductSpuService;
+import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
 
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
+import java.util.Collection;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
+import static cn.iocoder.yudao.module.product.dal.dataobject.category.ProductCategoryDO.PARENT_ID_NULL;
 import static cn.iocoder.yudao.module.product.enums.ErrorCodeConstants.*;
 
 /**
@@ -28,6 +35,9 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
 
     @Resource
     private ProductCategoryMapper productCategoryMapper;
+    @Resource
+    @Lazy // 循环依赖，避免报错
+    private ProductSpuService productSpuService;
 
     @Override
     public Long createCategory(ProductCategoryCreateReqVO createReqVO) {
@@ -61,14 +71,18 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
         if (productCategoryMapper.selectCountByParentId(id) > 0) {
             throw exception(CATEGORY_EXISTS_CHILDREN);
         }
-        // TODO 芋艿 补充只有不存在商品才可以删除
+        // 校验分类是否绑定了 SPU
+        Long spuCount = productSpuService.getSpuCountByCategoryId(id);
+        if (spuCount > 0) {
+            throw exception(CATEGORY_HAVE_BIND_SPU);
+        }
         // 删除
         productCategoryMapper.deleteById(id);
     }
 
     private void validateParentProductCategory(Long id) {
         // 如果是根分类，无需验证
-        if (Objects.equals(id, ProductCategoryDO.PARENT_ID_NULL)) {
+        if (Objects.equals(id, PARENT_ID_NULL)) {
             return;
         }
         // 父分类不存在
@@ -77,7 +91,7 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
             throw exception(CATEGORY_PARENT_NOT_EXISTS);
         }
         // 父分类不能是二级分类
-        if (Objects.equals(id, ProductCategoryDO.PARENT_ID_NULL)) {
+        if (!Objects.equals(category.getParentId(), PARENT_ID_NULL)) {
             throw exception(CATEGORY_PARENT_NOT_FIRST_LEVEL);
         }
     }
@@ -87,6 +101,26 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
         if (category == null) {
             throw exception(CATEGORY_NOT_EXISTS);
         }
+    }
+
+    @Override
+    public void validateCategoryList(Collection<Long> ids) {
+        if (CollUtil.isEmpty(ids)) {
+            return;
+        }
+        // 获得商品分类信息
+        List<ProductCategoryDO> list = productCategoryMapper.selectBatchIds(ids);
+        Map<Long, ProductCategoryDO> categoryMap = CollectionUtils.convertMap(list, ProductCategoryDO::getId);
+        // 校验
+        ids.forEach(id -> {
+            ProductCategoryDO category = categoryMap.get(id);
+            if (category == null) {
+                throw exception(CATEGORY_NOT_EXISTS);
+            }
+            if (!CommonStatusEnum.ENABLE.getStatus().equals(category.getStatus())) {
+                throw exception(CATEGORY_DISABLED, category.getName());
+            }
+        });
     }
 
     @Override
@@ -107,15 +141,16 @@ public class ProductCategoryServiceImpl implements ProductCategoryService {
 
     @Override
     public Integer getCategoryLevel(Long id) {
-        if (Objects.equals(id, ProductCategoryDO.PARENT_ID_NULL)) {
+        if (Objects.equals(id, PARENT_ID_NULL)) {
             return 0;
         }
         int level = 1;
-        for (int i = 0; i < 100; i++) {
-            ProductCategoryDO category = productCategoryMapper.selectById(id);
+        // for 的原因，是因为避免脏数据，导致可能的死循环。一般不会超过 100 层哈
+        for (int i = 0; i < Byte.MAX_VALUE; i++) {
             // 如果没有父节点，break 结束
+            ProductCategoryDO category = productCategoryMapper.selectById(id);
             if (category == null
-                    || Objects.equals(category.getParentId(), ProductCategoryDO.PARENT_ID_NULL)) {
+                    || Objects.equals(category.getParentId(), PARENT_ID_NULL)) {
                 break;
             }
             // 继续递归父节点
